refactor: 홈 컨트롤러 어노테이션 기반 컨트롤러 적용
