+++ jwp-basic-gradle/src/main/java/next/controller/HomeController.java
@@ -1,19 +1,22 @@
 package next.controller;
 
 import core.annotation.Controller;
+import core.annotation.RequestMapping;
+import core.annotation.RequestMethod;
 import core.mvcframework.ModelAndView;
+import core.mvcframework.controller.AbstractAnnotationController;
 import core.mvcframework.controller.AbstractController;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import next.service.QuestionService;
 
 @Controller
-public class HomeController extends AbstractController {
+public class HomeController extends AbstractAnnotationController {
 
     private final QuestionService questionService = new QuestionService();
 
-    @Override
-    public ModelAndView execute(final HttpServletRequest request, final HttpServletResponse response) {
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public ModelAndView home(final HttpServletRequest request, final HttpServletResponse response) {
         return jspView("home").addObject("questions", questionService.findAllOrderByCreatedDate());
     }
